Repareer fotoalbum, geef async modules een naam
