+++ resources/assets/js/extern.js
@@ -2,11 +2,8 @@ import $ from 'jquery';
 
 window.$ = window.jQuery = $;
 
-// jgallery is op page load nodig
-require('jgallery/dist/js/jgallery');
-
 $(function () {
 	$('body').removeClass('is-loading');
 
-	import('./extern-defer');
+	import(/* webpackChunkName: "extern-defer" */ './extern-defer');
 });
